requirements.txt
openai
pydantic
python-dotenv
streamlit>=1.28